Rename broadcastProgrammePresenter to CollapsedBroadcastProgrammePresenter
