+++ src/Ds2013/Organism/Programme/CollapsedBroadcastSubPresenters/CollapsedBroadcastProgrammeImagePresenter.php
@@ -1,15 +1,15 @@
 <?php
 declare(strict_types = 1);
-namespace App\Ds2013\Organism\Programme\BroadcastSubPresenters;
+namespace App\Ds2013\Organism\Programme\CollapsedBroadcastSubPresenters;
 
-use App\Ds2013\Organism\Programme\SubPresenters\ProgrammeBodyPresenter;
+use App\Ds2013\Organism\Programme\SubPresenters\ProgrammeImagePresenter;
 use App\Ds2013\Helpers\PlayTranslationsHelper;
 use App\Ds2013\Helpers\LiveBroadcastHelper;
 use BBC\ProgrammesPagesService\Domain\Entity\CollapsedBroadcast;
 use BBC\ProgrammesPagesService\Domain\Entity\Programme;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 
-class BroadcastProgrammeBodyPresenter extends ProgrammeBodyPresenter
+class CollapsedBroadcastProgrammeImagePresenter extends ProgrammeImagePresenter
 {
     /** @var CollapsedBroadcast */
     private $collapsedBroadcast;
@@ -35,35 +35,9 @@ class BroadcastProgrammeBodyPresenter extends ProgrammeBodyPresenter
         return $this->liveBroadcastHelper->simulcastUrl($this->collapsedBroadcast, $this->options['context_service']);
     }
 
-    public function isRepeat(): bool
-    {
-        return $this->collapsedBroadcast->isRepeat();
-    }
-
     public function isWatchableLive(): bool
     {
         return $this->liveBroadcastHelper->isWatchableLive($this->collapsedBroadcast, $this->options['advanced_live']);
-    }
-
-    public function isWorldServiceForeignLanguage(): bool
-    {
-        $network = $this->programme->getNetwork();
-        if ($network && $network->isWorldServiceInternational() && (string) $network->getNid() !== 'bbc_world_service') {
-            return true;
-        }
-        return false;
-    }
-
-    public function rewindUrl(): string
-    {
-        $url = $this->liveBroadcastHelper->simulcastUrl($this->collapsedBroadcast, $this->options['context_service']);
-        $url .= ((strpos($url, '?') !== false) ? '&' : '?') . 'rewindTo=current';
-        return $url;
-    }
-
-    public function translatePlayFromStart(): string
-    {
-        return $this->playTranslationsHelper->translatePlayFromStart($this->programme, $this->options['context_service']);
     }
 
     public function translatePlayLive(): string
@@ -73,6 +47,6 @@ class BroadcastProgrammeBodyPresenter extends ProgrammeBodyPresenter
 
     public function getTemplateVariableName(): string
     {
-        return 'programme_body';
+        return 'programme_image';
     }
 }
